2025/09/05/Gemini API and Python - Intro Tutorial/gemini_example/pyproject.toml
[project]
name = "gemini-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.33.0",
]
